% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Soundcheck Shiny App}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny]{shinyApp}}}
}
\value{
A Shiny app object
}
\description{
Launch the data validation Shiny application.
}
\examples{
\dontrun{
run_app()
}
}
