% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcheck_app.R
\name{soundcheck_app}
\alias{soundcheck_app}
\title{Launch the \code{soundcheck} app}
\usage{
soundcheck_app(settings, wav_folder = "wav", labels_file = "labels.csv")
}
\arguments{
\item{settings}{A \code{soundcheck} settings list, generated by \code{soundcheck_settings()}
(see the documentation for that function).}

\item{wav_folder}{The path (relative or absolute) to your folder of \code{.wav} files.
The default is a subfolder named \code{wav} within your working directory.}

\item{labels_file}{The filepath (relative or absolute) to a \code{.csv} of label results.
The default is \code{labels.csv} within your working directory.
If this file does not yet exist, the function will automatically create this file
based upon your settings.}
}
\value{
A \code{Shiny} app will open, displaying the spectrogram of the first recording
in your \code{wav_folder}. You can then navigate to other sound files as needed.
}
\description{
This app allows you to analyze a folder of \code{.wav} files: review spectrograms,
playback the sound, and label the contents of each file using your own custom categories.
}
\details{
The basic workflow is:
\enumerate{
\item (Make sure you have prepared your \code{settings} argument using the \code{soundcheck_settings()} function.
\item Run this function command to launch the Shiny app, which will open a browser window.
In the top center of the window, you will see a spectrogram of your first sound file
(it may take a few seconds to render).
\item Specify the analyst name in the top-left of the app.
\item Adjust spectrogram display as you need, using the slider settings.
Definitions for these settings are provided in the documentaiton for the \code{Spectrogram()} function.
Note that updating the spectrogram may take several seconds,
depending on the sample rate and duration of your recording.
If your sound files all have the same parameters (same sample rate,
similar frequency range of interest, and similar duration),
it may be beneficial to use the slider settings to optimize the spectrogram display,
\emph{then} revise the spectrogram parameters you specified in your \code{soundcheck_settings()} call, \emph{then} re-launch the app.
Doing so would save you the trouble of re-optimizing those slider settings
every time you (or a colleague) close and re-open the app while working on the same project.
Finally, note that the width of the displayed spectrogram may depend on the duration of your sound file;
files less than 30 seconds in duration will be scaled according to their duration,
so that short files are not too distorted by the aspect ratio of the display.
\item Listen to the recording (or a certain part of it), if you wish,
by double-clicking anywhere on the spectrogram. (To listen to the whole thing,
click on the far left of the spectrogram.)
Double-click again to stop playback. Adjust playback speed if you want, using the slider settings.
\item Below the spectrogram, you will see the label categories you specified in your \code{soundcheck_settings()} call.
Use those menus to label the features (or lack of features) in this sound file
\item To save your labels to your \code{labels_file}, click "Save & Next".
(To navigate to other files without saving your labels,
you can use the "Skip" or "Back" buttons, or manually select the file of your choice.).
Your labels will \emph{only} be saved when you click "Save & Next" --
labels will \emph{not} be saved when you click "Skip" or "Back".
\item If you are reviewing a file that already has a label record in the \code{labels_file},
those previous labels will be printed
in a table for the analyst to review. You can inspect the entire contents
of your \code{labels_file} using the table printed at the bottom of the app
(you may need to scroll down to find this table).
\item To filter the \code{.wav} files within your \code{wav_folder} --
e.g., to only the recordings you have not yet labeled, or only the recordings that you \emph{have} already labeled --
you can use the "Filter" dropdown menu (top-left in app).
\item When your work session is over, simply close the browser window.
If you modified the spectrogram settings during your work session,
take note of those changes in a notebook (they will be reset when you re-launch the app) --
or, better yet, modify the corresponding settings in your \code{soundcheck_settings()} call,
so that the app will re-launch for your next session using the settings you have been optimizing while you work.
}

When you are working on a large batch of sound files, the typical procedure will be:
(1) launch the app; (2) set your analyst name; (3) change the "Filter" menu to
'Not yet labeled'; (4) resume labeling.

Note that this app is optimized for \code{.wav} files with sampling rates of 44.1 kHz or lower, 1 minute or less,
but you may be able to adjust the spectrogram settings to make your files work.
}
