% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcheck_settings.R
\name{soundcheck_settings}
\alias{soundcheck_settings}
\title{Stage settings for SoundCheck app}
\usage{
soundcheck_settings(
  analysts = c("Ben", "Eric", "Other"),
  labels = list(Target_species = c("Not present", "Present", "Not sure"),
    Anthropogenic_noise = c("Not present", "Present", "Not sure")),
  na_default = TRUE,
  frequency_min = 0,
  frequency_max = 4000,
  window_length = 5,
  frequency_resolution = NULL,
  overlap = NULL,
  timestep_size = NULL,
  dynamic_range = NULL,
  window_type = NULL,
  window_parameter = NULL
)
}
\arguments{
\item{analysts}{desc}

\item{labels}{desc}

\item{na_default}{desc}

\item{frequency_min}{desc}

\item{frequency_max}{desc}

\item{window_length}{desc}

\item{frequency_resolution}{desc}

\item{overlap}{desc}

\item{timestep_size}{desc}

\item{dynamic_range}{desc}

\item{window_type}{desc}

\item{window_parameter}{desc}
}
\value{

}
\description{
Stage settings for SoundCheck app
}
